requirements.txt
Django>=3.0,<3.1
django-registration==3.1
django-debug-toolbar==2.2
psycopg2==2.8.5
django-countries==6.1.2
django-widget-tweaks==1.4.8
django-filter==2.2.0